Move NewList to ListView dir
